+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Imp/EntitiesSpec.hs
@@ -85,7 +85,7 @@ spec = describe "ENTITIES" $ do
     legacyTx <- switchToLegacyMode tx
     submitFailingTx
       legacyTx
-      [ injectFailure . IncompleteWithdrawals @era $
+      [ injectFailure . InexactWithdrawalsInLegacy @era $
           NE.singleton account1 $
             Mismatch lessThanReward1 reward1
       ]
@@ -205,14 +205,14 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       tx
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch (topAmount <+> subAmount) reward)]
       ]
     legacyTx <- switchToLegacyMode tx
     submitFailingTx
       legacyTx
-      [ injectFailure . IncompleteWithdrawals @era $
+      [ injectFailure . InexactWithdrawalsInLegacy @era $
           NE.singleton account $
             Mismatch topAmount (reward <-> subAmount)
       ]
@@ -231,14 +231,14 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       tx
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch (subAmount1 <+> subAmount2) reward)]
       ]
     legacyTx <- switchToLegacyMode tx
     submitFailingTx
       legacyTx
-      [ injectFailure . ExceededBalancesInWithdrawals @era $
+      [ injectFailure . WithdrawalsExceedAccountBalance @era $
           NE.singleton account $
             Mismatch (subAmount1 <+> subAmount2) reward
       ]
@@ -257,18 +257,18 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       subTxOverdraws
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch (atMostReward <+> moreThanReward) reward)]
       ]
     legacySubTxOverdraws <- switchToLegacyMode subTxOverdraws
     submitFailingTx
       legacySubTxOverdraws
-      [ injectFailure . IncompleteWithdrawals @era $
+      [ injectFailure . InexactWithdrawalsInLegacy @era $
           NE.singleton account $
             Mismatch atMostReward zero
       , injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch moreThanReward reward)]
       ]
@@ -281,7 +281,7 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       topTxOverdraws
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap
                 [(account, Mismatch (atMostReward <+> moreThanReward) reward)]
@@ -289,7 +289,7 @@ spec = describe "ENTITIES" $ do
     legacyTopTxOverdraws <- switchToLegacyMode topTxOverdraws
     submitFailingTx
       legacyTopTxOverdraws
-      [ injectFailure . IncompleteWithdrawals @era $
+      [ injectFailure . InexactWithdrawalsInLegacy @era $
           NE.singleton account $
             Mismatch moreThanReward (reward <-> atMostReward)
       ]
@@ -315,7 +315,7 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       tx
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch depositAmount zero)]
       ]
@@ -324,7 +324,7 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       legacyTx
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch depositAmount zero)]
       ]
@@ -345,7 +345,7 @@ spec = describe "ENTITIES" $ do
     submitFailingTx
       tx
       [ injectFailure $
-          ExceededBalancesInWithdrawals @era $
+          WithdrawalsExceedAccountBalance @era $
             fromJust $
               NE.fromMap [(account, Mismatch depositAmount zero)]
       ]