+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.4.0.0
 
+* Rename `EntitiesPredFailure` constructors:
+  - `IncompleteWithdrawals` to `InexactWithdrawalsInLegacy`
+  - `ExceededBalancesInWithdrawals` to `WithdrawalsExceedAccountBalance`
 * Remove `WithdrawalsExceedAccountBalance` constructor from `DijkstraUtxoPredFailure`
 * Add `MissingOriginalAccountsInWithdrawals` constructor to `EntitiesPredFailure`
 * Add `validateMissingOriginalAccountsInWithdrawals`, `validateMissingAccountsInWithdrawals`